+++ src/dawn/tests/end2end/FirstIndexOffsetTests.cpp
@@ -87,7 +87,8 @@ class FirstIndexOffsetTests : public DawnTest {
 
   private:
     wgpu::Buffer CreateVertexBuffer(uint32_t firstVertexOffset) {
-        std::vector<float> vertexData(firstVertexOffset * kComponentsPerVertex);
+        std::vector<float> vertexData(static_cast<size_t>(firstVertexOffset) *
+                                      kComponentsPerVertex);
         vertexData.insert(vertexData.end(), {0, 0, 0, 1});
         vertexData.insert(vertexData.end(), {0, 0, 0, 1});
         return utils::CreateBufferFromData(device, vertexData.data(),
